requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
langchain-text-splitters>=0.0.1
gitpython>=3.1.40
PyGithub>=2.1.1
pymupdf>=1.23.0
chromadb>=0.4.22
faiss-cpu>=1.7.4
google-genai>=1.0.0
fastembed>=0.2.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
numpy
